requirements.txt
# Flask Core
Flask==3.0.2
Flask-SQLAlchemy==3.1.1
Flask-Cors==4.0.0

# Auth
PyJWT==2.8.0
werkzeug==3.0.1

# Templates & Markdown
Jinja2==3.1.3
Markdown==3.5.2
MarkupSafe==2.1.5

# Utils
python-dotenv==1.0.0
beautifulsoup4
itsdangerous==2.1.2
click==8.1.7

# API Docs
flasgger==0.9.7.1
